全局异常处理

+++ src/main/java/xyz/esp8266/community/service/QuestionService.java
@@ -14,8 +14,6 @@ import xyz.esp8266.community.mapper.UserMapper;
 import xyz.esp8266.community.model.Question;
 import xyz.esp8266.community.model.QuestionExample;
 import xyz.esp8266.community.model.User;
-
-import javax.swing.undo.CannotUndoException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -74,7 +72,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public PaginationDTO list(Integer userId, Integer page, Integer size) {
+    public PaginationDTO list(Long userId, Integer page, Integer size) {
         PaginationDTO paginationDTO = new PaginationDTO();
         QuestionExample questionExample = new QuestionExample();
         questionExample.createCriteria().andCreatorEqualTo(userId);
@@ -103,7 +101,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question = questionMapper.selectByPrimaryKey(id);
         if(question == null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -115,7 +113,7 @@ public class QuestionService {
         return questionDTO;
     }
     // 增加阅读数
-    public void incView(Integer id) {
+    public void incView(Long id) {
         Question question = new Question();
         question.setId(id);
         question.setViewCount(1);
